# A nice speed up of 3D points clouds by ~69%