--- a/.cargo/config.toml
+++ b/.cargo/config.toml
@@ -2,6 +2,9 @@
 # `cargo rerun` is short a convenient shorthand, skipping the web viewer.
 rerun = "run --package rerun-cli --no-default-features --features native_viewer --"
 
+# like `rerun`, but with --release
+rerun-release = "run --package rerun-cli --no-default-features --features native_viewer --release --"
+
 # `cargo rerun-web` is short a convenient shorthand for building & starting the web viewer.
 rerun-web = "run --package rerun-cli --no-default-features --features web_viewer -- --web-viewer"