--- a/crates/re_space_view_spatial/Cargo.toml
+++ b/crates/re_space_view_spatial/Cargo.toml
@@ -42,6 +42,7 @@ itertools.workspace = true
 macaw = { workspace = true, features = ["with_serde"] }
 nohash-hasher.workspace = true
 parking_lot.workspace = true
+rayon.workspace = true
 serde = "1"
 smallvec = { workspace = true, features = ["serde"] }
 vec1.workspace = true