--- a/crates/re_space_view_spatial/src/contexts/annotation_context.rs
+++ b/crates/re_space_view_spatial/src/contexts/annotation_context.rs
@@ -14,6 +14,7 @@ impl ViewContextSystem for AnnotationSceneContext {
         ctx: &mut re_viewer_context::ViewerContext<'_>,
         query: &re_viewer_context::ViewQuery<'_>,
     ) {
+        re_tracing::profile_function!();
         self.0.load(
             ctx,
             &query.latest_at_query(),