--- a/crates/re_space_view_spatial/src/parts/points3d.rs
+++ b/crates/re_space_view_spatial/src/parts/points3d.rs
@@ -172,7 +172,7 @@ impl Points3DPart {
             }
         }
 
-        load_keypoint_connections(ent_context, ent_path, keypoints);
+        load_keypoint_connections(ent_context, ent_path, &keypoints);
 
         self.data.extend_bounding_box_with_points(
             arch_view