--- a/crates/re_viewer_context/src/annotations.rs
+++ b/crates/re_viewer_context/src/annotations.rs
@@ -81,7 +81,7 @@ impl From<ClassDescription> for CachedClassDescription {
     }
 }
 
-#[derive(Clone, Debug)]
+#[derive(Clone, Copy, Debug)]
 pub struct ResolvedClassDescription<'a> {
     pub class_id: Option<ClassId>,
     pub class_description: Option<&'a ClassDescription>,