### What
On native, use `rayon` to parallelize the gathering of positions, radii, colors, and picking ids. Of course, this only works on native. I also noticed that we spent a lot of time generating identical annotation infos for all the points, so I happy-pathed that.

Test: `cargo rerun-release ../opf.rrd --profile`, looking at just the 3D view, averaged over ~30 frames:

Full frame 37.0ms -> 26.6ms ≈ 39% faster

`Points3DPart`: 25.7ms -> 15.2ms ≈ 69% faster

Before:
![Screenshot 2023-08-27 at 17 39 52](https://github.com/rerun-io/rerun/assets/1148717/95b2de1b-ab21-489b-8e05-039e8de50692)

After:
![image](https://github.com/rerun-io/rerun/assets/1148717/2e0f7dc7-ccce-4cd8-9157-162bb642c795)



### Checklist
* [x] I have read and agree to [Contributor Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and the [Code of Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [x] I've included a screenshot or gif (if applicable)
* [x] I have tested [demo.rerun.io](https://demo.rerun.io/pr/3114/examples/structure_from_motion/) (if applicable)

- [PR Build Summary](https://build.rerun.io/pr/3114)
- [Docs preview](https://rerun.io/preview/d70a51a580149f974aec0c089fb8f899e132c518/docs) <!--DOCS-PREVIEW-->
- [Examples preview](https://rerun.io/preview/d70a51a580149f974aec0c089fb8f899e132c518/examples) <!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW--><!--EXAMPLES-PREVIEW-->
- [Recent benchmark results](https://ref.rerun.io/dev/bench/)
- [Wasm size tracking](https://ref.rerun.io/dev/sizes/)