@emilk 2023-08-27
Zooming in on what's slow now, it is obvious that we need to optimize the Color deserializer:

![image](https://github.com/rerun-io/rerun/assets/1148717/7dd34190-45dc-4b80-a1fe-c9115c6013cc)

* https://github.com/rerun-io/rerun/issues/3109